+++ service-go/build.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+set -x
+
+docker build -t interview-service .
+
+set +x
