+++ service-go/cmd/service/service.go
@@ -1,4 +1,4 @@
-package service
+package main
 
 import (
 	"fmt"
@@ -9,11 +9,40 @@ import (
 	"interview-service/internal/api"
 	"interview-service/internal/api/interview"
 
+	"flag"
+	"os"
+
+	"github.com/joho/godotenv"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/reflection"
 )
 
-func Start() {
+func main() {
+	loglevel := flag.Int("logLevel", 4, "Useful Log levels: Warn = 3; Info = 4; Debug = 5;")
+	logGrpc := flag.Bool("logGrpc", false, "Turn ON/OFF grpc middleware logs")
+	flag.Parse()
+
+	log.SetFormatter(&log.JSONFormatter{})
+	log.SetOutput(os.Stdout)
+
+	if *loglevel >= 0 && *loglevel <= 6 {
+		log.SetLevel(log.Level(*loglevel))
+	}
+
+	if *logGrpc {
+		err := godotenv.Load("logGrpc.env")
+		if err != nil {
+			log.Infof("Error loading .env file: %+v", err)
+		}
+		log.Debugf("Loaded .env file")
+		log.Debugf("Severity: %s", os.Getenv("GRPC_GO_LOG_SEVERITY_LEVEL"))
+		log.Debugf("Verbosity: %s", os.Getenv("GRPC_GO_LOG_VERBOSITY_LEVEL"))
+	}
+
+	start()
+}
+
+func start() {
 	address := fmt.Sprintf("localhost:%d", 8080)
 	log.Infof("Starting interview service at %s", address)
